+++ cmd/cli/cmd/alpha/manifest-gen/implementation/helm.go
@@ -15,7 +15,7 @@ func NewHelm() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "helm [MANIFEST_PATH] [HELM_CHART_NAME]",
 		Short: "Generate Helm chart based manifests",
-		Long:  "Generate Helm based manifests based on a Helm chart",
+		Long:  "Generate Implementation manifests based on a Helm chart",
 		Args: func(cmd *cobra.Command, args []string) error {
 			if len(args) != 2 {
 				return errors.New("accepts two arguments: [MANIFEST_PATH] [HELM_CHART_NAME]")
@@ -57,8 +57,8 @@ func NewHelm() *cobra.Command {
 
 	cmd.Flags().StringVarP(&helmCfg.InterfacePathWithRevision, "interface", "i", "", "Path with revision of the Interface, which is implemented by this Implementation")
 	cmd.Flags().StringVarP(&helmCfg.ManifestRevision, "revision", "r", "0.1.0", "Revision of the Implementation manifest")
-	cmd.Flags().StringVar(&helmCfg.RepoURL, "repo", "", "URL of the Helm repository")
-	cmd.Flags().StringVarP(&helmCfg.Version, "version", "v", "", "Version of the Helm chart")
+	cmd.Flags().StringVar(&helmCfg.ChartRepoURL, "repo", "", "URL of the Helm repository")
+	cmd.Flags().StringVarP(&helmCfg.ChartVersion, "version", "v", "", "Version of the Helm chart")
 
 	return cmd
 }
