+++ internal/cli/alpha/manifestgen/manifestgen.go
@@ -2,24 +2,11 @@ package manifestgen
 
 import (
 	"bytes"
-	"fmt"
-	"strings"
 	"text/template"
 
 	"github.com/Masterminds/sprig"
 	"github.com/pkg/errors"
 )
-
-// Config stores the generic input parameters for content generation
-type Config struct {
-	ManifestPath     string
-	ManifestRevision string
-}
-
-type templatingConfig struct {
-	Template string
-	Input    interface{}
-}
 
 func generateManifests(cfgs []*templatingConfig) ([]string, error) {
 	manifests := make([]string, 0, len(cfgs))
@@ -50,16 +37,4 @@ func generateManifest(cfg *templatingConfig) (string, error) {
 	}
 
 	return manifest.String(), nil
-}
-
-func splitPathToPrefixAndName(path string) (string, string, error) {
-	parts := strings.Split(path, ".")
-	if len(parts) < 3 {
-		return "", "", fmt.Errorf("manifest path must have prefix and name")
-	}
-
-	prefix := strings.Join(parts[2:len(parts)-1], ".")
-	name := parts[len(parts)-1]
-
-	return prefix, name, nil
 }
